requirements.txt
Flask>=2.3.2,<3.0
SQLAlchemy>=2.0.20,<3.0
psycopg2-binary>=2.9.7,<3.0
python-dotenv>=1.0.0,<2.0
